requirements.txt
pandas
python-dotenv
google-generativeai
google-auth-httplib2
google-api-core
google-api-python-client
google-generativeai
